jwt token check

+++ nmdrinkbook-api/controllers/article.controller.js
@@ -11,30 +11,34 @@ export const articleController = () => {
     
         var authHeader = req.headers.authorization
         
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
-
-        if (!is_admin) {
-            res.status(401).send()
-        } else {
-            const article = {
-                articleName: req.body.articleName,
-                description: req.body.description,
-                alcoholVolume: req.body.alcoholVolume,
-                createdBy: req.createdBy,
-                modifiedBy: req.modifiedBy
-            }
-        
-            Article.create(article)
-                .then(data => {
-                    res.send(data);
-                })
-                .catch(err => {
-                    res.status(500).send({
-                        message: err.message || "Unknown error."
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+    
+            if (!is_admin) {
+                res.status(401).send()
+            } else {
+                const article = {
+                    articleName: req.body.articleName,
+                    description: req.body.description,
+                    alcoholVolume: req.body.alcoholVolume,
+                    createdBy: req.createdBy,
+                    modifiedBy: req.modifiedBy
+                }
+            
+                Article.create(article)
+                    .then(data => {
+                        res.send(data);
                     })
-                });
+                    .catch(err => {
+                        res.status(500).send({
+                            message: err.message || "Unknown error."
+                        })
+                    });
+            }
+        } else {
+            res.status(401).send()
         }
     };
 
@@ -43,30 +47,34 @@ export const articleController = () => {
         req.body.modifiedBy = req.modifiedBy
         var authHeader = req.headers.authorization
         
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
-
-        if (!is_admin) {
-            res.status(401).send()
-        } else {
-            Article.update(req.body, {
-                where: { articleId: id }
-            })
-                .then(num => {
-                    if (num == 1) {
-                        res.send(req.body)
-                    } else {
-                        res.status(500).send({
-                            message: "error"
-                        })
-                    }
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+    
+            if (!is_admin) {
+                res.status(401).send()
+            } else {
+                Article.update(req.body, {
+                    where: { articleId: id }
                 })
-                .catch(err => {
-                    res.status(500).send({
-                        message: "error catch"
+                    .then(num => {
+                        if (num == 1) {
+                            res.send(req.body)
+                        } else {
+                            res.status(500).send({
+                                message: "error"
+                            })
+                        }
                     })
-                })
+                    .catch(err => {
+                        res.status(500).send({
+                            message: "error catch"
+                        })
+                    })
+            }
+        } else {
+            res.status(401).send()
         }
     }
 
@@ -74,30 +82,34 @@ export const articleController = () => {
         const id = req.params.id
         var authHeader = req.headers.authorization
         
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
-
-        if (!is_admin) {
-            res.status(401).send()
-        } else {
-            Article.destroy({
-                where: { articleId: id }
-            })
-                .then(num => {
-                    if (num == 1) {
-                        res.status(204).send();
-                    } else {
-                        res.send({
-                            message: `error`
-                        });
-                    }
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+    
+            if (!is_admin) {
+                res.status(401).send()
+            } else {
+                Article.destroy({
+                    where: { articleId: id }
                 })
-                .catch(err => {
-                    res.status(500).send({
-                    message: "error catch"
-                    });
-              });
+                    .then(num => {
+                        if (num == 1) {
+                            res.status(204).send();
+                        } else {
+                            res.send({
+                                message: `error`
+                            });
+                        }
+                    })
+                    .catch(err => {
+                        res.status(500).send({
+                        message: "error catch"
+                        });
+                  });
+            }
+        } else {
+            res.status(401).send()
         }
     }
