+++ nmdrinkbook-api/controllers/unit.controller.js
@@ -12,29 +12,33 @@ export const unitController = () => {
     
         var authHeader = req.headers.authorization
         
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
 
-        if (!is_admin) {
-            res.status(401).send()
-        } else {
-            const unit = {
-                symbol: req.body.symbol,
-                createdBy: req.createdBy,
-                modifiedBy: req.modifiedBy
-            }
-        
-            Unit.create(unit)
-                .then(data => {
-                    res.send(data);
-                    
-                })
-                .catch(err => {
-                    res.status(500).send({
-                        message: err.message || "Unknown error."
+            if (!is_admin) {
+                res.status(401).send()
+            } else {
+                const unit = {
+                    symbol: req.body.symbol,
+                    createdBy: req.createdBy,
+                    modifiedBy: req.modifiedBy
+                }
+            
+                Unit.create(unit)
+                    .then(data => {
+                        res.send(data);
+                        
                     })
-                });
+                    .catch(err => {
+                        res.status(500).send({
+                            message: err.message || "Unknown error."
+                        })
+                    });
+            }
+        } else {
+            res.status(401).send()
         }
     };
 
@@ -42,31 +46,33 @@ export const unitController = () => {
         const id = req.params.id
         req.body.modifiedBy = req.modifiedBy
         var authHeader = req.headers.authorization
-        
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
 
-        if (!is_admin) {
-            res.status(401).send();
-        } else {
-            Unit.update(req.body, {
-                where: { unitId: id }
-            })
-                .then(num => {
-                    if (num == 1) {
-                        res.send(req.body)
-                    } else {
-                        res.status(500).send({
-                            message: "error"
-                        })
-                    }
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+
+            if (!is_admin) {
+                res.status(401).send();
+            } else {
+                Unit.update(req.body, {
+                    where: { unitId: id }
                 })
-                .catch(err => {
-                    res.status(500).send({
-                        message: "error catch"
+                    .then(num => {
+                        if (num == 1) {
+                            res.send(req.body)
+                        } else {
+                            res.status(500).send({
+                                message: "error"
+                            })
+                        }
                     })
-                })
+                    .catch(err => {
+                        res.status(500).send({
+                            message: "error catch"
+                        })
+                    })
+            }
         }
     }
 
@@ -74,30 +80,34 @@ export const unitController = () => {
         const id = req.params.id
         var authHeader = req.headers.authorization
         
-        var token = authHeader.substring(7, authHeader.length)
-        var decoded = jwt_decode(token)
-        let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
-
-        if (!is_admin) {
-            res.status(401).send()
-        } else {
-            Unit.destroy({
-                where: { unitId: id }
-            })
-                .then(num => {
-                    if (num == 1) {
-                        res.status(204).send();
-                    } else {
-                        res.send({
-                            message: `error`
-                        });
-                    }
+        if (authHeader) {
+            var token = authHeader.substring(7, authHeader.length)
+            var decoded = jwt_decode(token)
+            let is_admin = decoded['resource_access']['nmclient']['roles'].includes('admin')
+    
+            if (!is_admin) {
+                res.status(401).send()
+            } else {
+                Unit.destroy({
+                    where: { unitId: id }
                 })
-                .catch(err => {
-                    res.status(500).send({
-                    message: "error catch"
-                    });
-              });
+                    .then(num => {
+                        if (num == 1) {
+                            res.status(204).send();
+                        } else {
+                            res.send({
+                                message: `error`
+                            });
+                        }
+                    })
+                    .catch(err => {
+                        res.status(500).send({
+                        message: "error catch"
+                        });
+                  });
+            }
+        } else {
+            res.status(401).send()
         }
     }
